requirements.txt
# Weather Edge — Dependencies
# Install: pip install -r requirements.txt
# Python 3.11+ required (for zoneinfo, modern type hints)

# Async HTTP client for API calls (Open-Meteo, NWS, Kalshi, Discord)
aiohttp~=3.9

# Environment variable management (.env files)
python-dotenv~=1.0

# RSA-PSS signing for Kalshi API authentication
cryptography~=42.0

# Vectorized KDE probability computation (~50x faster than pure Python)
numpy~=1.26

# Retry logic with exponential backoff (Kalshi API)
tenacity~=8.2

# Structured logging
structlog~=24.1

# Terminal formatting
rich~=13.7
